requirements.txt
Flask
Redis
pillow
tensorflow
cassandra-driver
